chore: move to util file
